separated sort logic for bundles
